StorageWriter recovery from writes with unknown outcome/bad offset writes.



This Pull Request adds functionality to the StorageWriter to make it automatically detect and reconcile operations with unknown outcomes (i.e., Timeouts) and other data corruption scenarios.

Causes

    Unknown outcome: Timeout during write or concat.
    Data corruption: Another actor writing data to segment.

Detection

    Detection usually happens not on the initial call (which times out), but on the next attempt to commit the operation (any storage error is retried by the Writer until it succeeds or until it detects that it cannot make any sort of progress).
    BadOffsetException from calls to Storage.write and Storage.concat. -- Both these operations take in an 'offset' as argument, and they ensure that the write only occurs if the current length of the Segment in Storage matches that offset.
    SegmentNotExistsException/BadOffsetException: after a merge was executed but timed out (if the concat happened, the source (transaction) segment is already deleted, so when a retry is attempted, either one of these exceptions might be raised by the Storage adapter).

System State

    After detection, we still have all unacknowledged operations in our operation queue (in the SegmentAggregator). -- If a Merge, then the Merge Op should be the first one. -- If a Flush, then the append(s) should still be there, at the head of the queue.

Resolution

    The SegmentAggregator that is involved enters the 'Reconciling' state (it has multiple states, including: Writing (normal operation mode), ReconciliationNeeded (when a possible corruption was detected - see above) and Reconciling (when it is actively is reconciling operations).
    Gets up-to-date information about the segment from Storage (RSL - Real StorageLength, RSS - Real Storage SealStatus)
    Performs initial validation:
        RSL == MSL (Metadata Storage Length): All is OK, no need to reconcile
        RSL < MSL: Data loss (unrecoverable)
        RSL > MSL: execute Reconciliation Procedure (see below)
        RSL > DLL (DurableLog length): impossible (unrecoverable)
    Performs Reconciliation Procedure, by processing each Operation in the queue until it finds one that exceeds the RSL (see above).
        Append: Read from Storage and verify, byte-by-byte, that the data matches
        Merge: Verify Source (Transaction) Segment is deleted and that the entire Transaction fits in the remaining Parent Segment length (Transactions can be pretty big, so we can't check them byte-by-byte)
        Seal: verify segment is sealed.
    Exit Reconcile Mode when an operation was encountered that has at least one byte beyond the RSL (see above).
        If any of the operations cannot be reconciled, the writer is shut down, which causes the entire container to shut down as well to prevent further data corruption.
        TBD: when such a corruption happens, we need to prevent the container from restarting on a different node. This means further work is required on the container side to prevent that.

+++ service/server/src/main/java/com/emc/pravega/service/server/mocks/LocalSegmentContainerManager.java
@@ -103,7 +103,7 @@ public class LocalSegmentContainerManager implements SegmentContainerManager {
 
     @Override
     public CompletableFuture<Void> initialize(Duration timeout) {
-        int traceId = LoggerHelpers.traceEnter(log, "initialize");
+        long traceId = LoggerHelpers.traceEnter(log, "initialize");
         ensureNotClosed();
         TimeoutTimer timer = new TimeoutTimer(timeout);
         ArrayList<CompletableFuture<Void>> futures = new ArrayList<>();
